% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstimates.R
\name{getASISest}
\alias{getASISest}
\title{"As is" estimates}
\usage{
getASISest(outcome, score)
}
\arguments{
\item{outcome}{Vector of binary outcome for each observation.}

\item{score}{Numeric vector of continuous predicted risk score.}
}
\value{
A data frame with 4 columns
(score, score percentile, outcome, estimate).
}
\description{
This function does no estimation, but uses the score as it is
(it works like an identity function).
}
\examples{
# Read in example data
auroc <- read.csv(system.file("extdata", "sample.csv", package = "stats4phc"))
rscore <- auroc$predicted
truth <- as.numeric(auroc$actual)

tail(getASISest(outcome = truth, score = rscore), 10)

}
\seealso{
\code{\link[=getBINNEDest]{getBINNEDest()}} \code{\link[=getCGAMest]{getCGAMest()}} \code{\link[=getGAMest]{getGAMest()}} \code{\link[=getMSPLINEest]{getMSPLINEest()}} \code{\link[=getPAVAest]{getPAVAest()}}
}
