% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstimates.R
\name{getBINNEDest}
\alias{getBINNEDest}
\title{Binned Risk Estimates}
\usage{
getBINNEDest(
  outcome,
  score,
  quantiles = NULL,
  bins = NULL,
  right = TRUE,
  errorbar.sem = NULL
)
}
\arguments{
\item{outcome}{Vector of binary outcome for each observation.}

\item{score}{Numeric vector of continuous predicted risk score.}

\item{quantiles}{Numeric; quantiles to split bins.}

\item{bins}{Numeric; number of evenly spaced bins or bin locations.}

\item{right}{Logical indicating right closed interval. Defaults to \code{TRUE}.}

\item{errorbar.sem}{Scalar numeric representing the number of standard error from the means
(SEM) used to calculate risk error bar.}
}
\value{
A data frame with 4 columns
(score, score percentile, outcome, estimate).
Additionally, there is an attribute "errorbar" holding the error-bar data if
\code{errorbar.sem} was specified.
}
\description{
Calculates bins based on number of evenly spaced bins or n-tiles.
Determines average risk within bins, used for risk estimates.
}
\examples{
# Read in example data
auroc <- read.csv(system.file("extdata", "sample.csv", package = "stats4phc"))
rscore <- auroc$predicted
truth <- as.numeric(auroc$actual)

getBINNEDest(outcome = truth, score = rscore)

}
\seealso{
\code{\link[=getASISest]{getASISest()}} \code{\link[=getCGAMest]{getCGAMest()}} \code{\link[=getGAMest]{getGAMest()}} \code{\link[=getMSPLINEest]{getMSPLINEest()}} \code{\link[=getPAVAest]{getPAVAest()}}
}
