% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstimates.R
\name{getGAMest}
\alias{getGAMest}
\title{GAM Risk Estimates}
\usage{
getGAMest(
  outcome,
  score,
  k = -1,
  bs = "tp",
  method = "REML",
  logscores = FALSE,
  fitonPerc = TRUE
)
}
\arguments{
\item{outcome}{Vector of binary outcome for each observation.}

\item{score}{Numeric vector of continuous predicted risk score.}

\item{k}{Numeric to specify the upper limit of basis functions to fit for GAM.
See \code{\link[mgcv:s]{mgcv::s()}} for more details. Defaults to -1.}

\item{bs}{Character string to specify spline type.
See \code{\link[mgcv:s]{mgcv::s()}} for more details. Defaults to \code{"tp"}.}

\item{method}{Character string to specify method type.
See \code{\link[mgcv:s]{mgcv::s()}} for more details. Defaults to "REML".}

\item{logscores}{Logical; if \code{TRUE}, fit gam on log scores. Defaults to \code{FALSE}.}

\item{fitonPerc}{Logical; if \code{TRUE}, fit gam on risk percentiles. Defaults to \code{TRUE}.}
}
\value{
A data frame with 4 columns
(score, score percentile, outcome, estimate).
}
\description{
Fits a Generalized Additive Model to estimate risk, given a vector of binary outcome,
and a vector of scores.
}
\examples{
# Read in example data
auroc <- read.csv(system.file("extdata", "sample.csv", package = "stats4phc"))
rscore <- auroc$predicted
truth <- as.numeric(auroc$actual)

tail(getGAMest(outcome = truth, score = rscore), 10)

}
\seealso{
\code{\link[=getASISest]{getASISest()}} \code{\link[=getBINNEDest]{getBINNEDest()}} \code{\link[=getCGAMest]{getCGAMest()}} \code{\link[=getMSPLINEest]{getMSPLINEest()}} \code{\link[=getPAVAest]{getPAVAest()}}
}
