% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstimates.R
\name{getMSPLINEest}
\alias{getMSPLINEest}
\title{Monotone Spline Risk Estimates}
\usage{
getMSPLINEest(outcome, score, k = 10, fitonPerc = TRUE)
}
\arguments{
\item{outcome}{Vector of binary outcome for each observation.}

\item{score}{Numeric vector of continuous predicted risk score.}

\item{k}{Numeric to specify the upper limit of basis functions to fit for GAM.
See \code{\link[mgcv:s]{mgcv::s()}} for more details. Defaults to -1.}

\item{fitonPerc}{Logical; if \code{TRUE}, fit gam on risk percentiles. Defaults to \code{TRUE}.}
}
\value{
A data frame with 4 columns
(score, score percentile, outcome, estimate).
}
\description{
Fits a Monotone constrained Generalized Additive Model (GAM) to estimate risk,
given a vector of binary outcomes and a vector of scores.
}
\examples{
# Read in example data
auroc <- read.csv(system.file("extdata", "sample.csv", package = "stats4phc"))
rscore <- auroc$predicted
truth <- as.numeric(auroc$actual)

tail(getMSPLINEest(outcome = truth, score = rscore), 10)

}
\seealso{
\code{\link[=getASISest]{getASISest()}} \code{\link[=getBINNEDest]{getBINNEDest()}} \code{\link[=getCGAMest]{getCGAMest()}} \code{\link[=getGAMest]{getGAMest()}} \code{\link[=getPAVAest]{getPAVAest()}}
}
