% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats4phc-package.R
\docType{package}
\name{stats4phc-package}
\alias{stats4phc}
\alias{stats4phc-package}
\title{stats4phc: Performance evaluation for the prognostic value of predictive models intended to
support personalized healthcare through predictiveness curves and positive / negative
predictive values}
\description{
vignette("stats4phc.Rmd")
}
\seealso{
Useful links:
\itemize{
  \item \url{https://genentech.github.io/stats4phc}
}

}
\author{
\strong{Maintainer}: Ondrej Slama \email{ondrej.slama@roche.com}

Authors:
\itemize{
  \item Darrick Shen \email{shend9@gene.com}
  \item Verena Steffen \email{steffen.verena@gene.com}
  \item Doug Kelkhoff \email{doug.kelkhoff@roche.com}
  \item Michel Friesenhahn \email{friesem1@gene.com}
  \item Christina Rabe \email{rabec@gene.com}
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
